[cita chain] refactor errors
