--- cita-chain/types/src/errors/error.rs
@@ -1,41 +1,0 @@
-// Copyright 2015-2017 Parity Technologies (UK) Ltd.
-// This file is part of Parity.
-
-// This software is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This software is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
-
-//! General error types for use in ethcore.
-
-use std::fmt;
-
-use super::execution::ExecutionError;
-
-#[derive(Debug)]
-pub enum Error {
-    /// Error concerning EVM code execution.
-    Execution(ExecutionError),
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Error::Execution(ref err) => err.fmt(f),
-        }
-    }
-}
-
-impl From<ExecutionError> for Error {
-    fn from(err: ExecutionError) -> Error {
-        Error::Execution(err)
-    }
-}
